change Mission 1 round
